+++ refinedstorage2-platform-api/src/main/java/com/refinedmods/refinedstorage2/platform/api/storage/channel/PlatformStorageChannelType.java
@@ -42,4 +42,6 @@ public interface PlatformStorageChannelType<T> extends StorageChannelType<T> {
     int getXTexture();
 
     int getYTexture();
+
+    long normalizeAmount(double amount);
 }